Finished Project

+++ app/Http/Controllers/TagsController.php
@@ -16,7 +16,6 @@ class TagsController extends Controller
    */
   public function index()
   {
-    //
     return view('tags.index')->with('tags', Tag::all());
   }
 
@@ -99,6 +98,11 @@ class TagsController extends Controller
    */
   public function destroy(Tag $tag)
   {
+
+    if ($tag->posts->count() > 0) {
+      session()->flash('error', 'Tag cannot be deleted, because it is associated to some posts.');
+      return redirect()->back();
+    }
     $tag->delete();
 
     session()->flash('success', 'Tag deleted successfully.');
